Refactor base package
- Move `FeedbackActionDispatcher` and related components into base package
- Refactor `FeedbackAwareModelCommand`. Remove `ModelRootChangeListener` API from this command and add it to the command stack instead. This ensures that the model listeners are notified after the update has been processed succeessfully. It also   makes it easier to rebind the `UpdateModelCommand` as you don't have to consider the listeners anymore.
