+++ packages/client/src/features/layout/layout-elements-action.spec.ts
@@ -41,9 +41,9 @@ import {
     TYPES,
     defaultModule
 } from '~glsp-sprotty';
+import { FeedbackActionDispatcher } from '../../base/feedback/feedback-action-dispatcher';
 import { resizeFeature } from '../change-bounds/model';
 import { SelectionService } from '../select/selection-service';
-import { FeedbackActionDispatcher } from '../tool-feedback/feedback-action-dispatcher';
 import {
     AlignElementsAction,
     AlignElementsActionHandler,
